requirements.txt
colorama
openai
python-dotenv
requests
beautifulsoup4
selenium
webdriver_manager
yt-dlp
ffmpeg-python
elevenlabs
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
imageio>=2.31.1
librosa>=0.10.1
hydra-core>=1.3.2
moviepy>=1.0.3
numpy==1.24.3
torch
transformers>=4.35.2
openai-whisper
aiohttp